backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
httpx
pytest
pytest-asyncio
